rebalancer-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
ib_async==2.0.1
aiohttp==3.9.1
asyncio-throttle==1.0.2
PyYAML==6.0.1
nest-asyncio==1.5.8